Fix animited_box/integrated_main
Sometime the plugin is not loaded when running integrated_main.
Whether it's loaded or not seems to be random. The reason is World::Step() only call LoadPlugins() if SensorsInitialized().
However, dataPtr->sensorsInitialized variable is not initialized and its value can be true or false.
gzserver calls sensors::run_once() to initialize sensors (through Server::Run()), so it does not have this issue.
To fix this problem, we first initialize sensorsInitialized correctly and add sensors::run_once() to integrated_main.cc