Implemented Save Canvas Layout
Added Option to save / load canvas layout to the plot plugin.
Based on new "AddVariable" due to some inconsistency when manually adding plots to the canvas. Uses same mechanism as "OnAddVariable", except there is no sender but the latest IncrementalPlot is being used from the list that can be retrieved by "Plots"