WideAngleCamera: specify cube map texture format
The WideAngleCamera has a texture for a cube map that
uses the same pixel format as the camera image. If there
are intermediate shaders applied to the camera, the choice
of this image format may cause pixel values to be truncated.
To provide flexibility to the user, the pixel format for
the cube map texture can be specified with a custom SDFormat
element in `//camera/lens/gazebo:env_texture_format`
using the same values as `//camera/image/format`.

Add friend WideAngleCamera statement to Camera.hh to allow
access to private OgrePixelFormat method.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>