Add models and world to show distortion bug
There is a bug in the initialization of cameras with distortion.
Loading multiple cameras with different distortion parameters
at the same time causes some of the same distortion parameters
to be applied to all those cameras. Spawning a model with a single
distorted camera afterwards works fine, even in a world that
already contains distorted cameras.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>