ODEJoint: don't SetStiffnessDamping for gearbox
The Gearbox joint does not support the stiffness and
damping parameters, so exit ODEJoint::SetStiffnessDamping
early if for Gearbox joints.

Without this change, simulations including a gearbox joint
stream a prohibitive amount of unnecessary messages to a
log file with complaints about APIs that are unsupported
by the Gearbox joint.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>