Support plotting for entities with / in the name
I noticed while testing the rexrov models from the rexrov
model in Field-Robotics-Lab/dave that the plotting window
does not work properly when elements of a model include
a forward-slash `/` in the entity name. To fix the problem,
the `/` characters are encoded as %2f in physics::Base::URI().
Corresponding decoding is added to the gui/plot/Palette class.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>