Parse ode_quiet physics parameter from SDFormat (#3194)
The ODE world-step solver has a tendency to spam the console
with LCP Internal Error messages. We added the `ode_quiet`
parameter to the `ODEPhysics::SetParam` C++ API to disable
these messages, but it would be more convenient to set
`ode_quiet` in an SDFormat world file. The PresetManager
does pass `ode_quiet` to the C++ API, but there is a casting
error. Since the ode_quiet element is not part of the SDFormat
spec, it is encoded as a string in the PresetManager.
Parsing it as a string using an sdf::Param object fixes it.
A test is added using world_step.world to confirm the fix.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>