Support shininess value for each Visual in a Model (#3235)
This moves the parsing of shininess values from
World::LoadModel to Link::UpdateVisualMsg and
stores the values by scoped Visual name rather
than by Model name. The per-model gz services
are also replaced by a single /shininess
service. A World::ShininessByScopedName method
for accessing the shininess map is provided, and
a mutex is used to protect access to the map.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>