Fix wide-angle camera lensflare occlusion (#3276)
* Fixing Z check in WideAngleCamera::Project3d

Signed-off-by: Sanjuksha Nirgude <sanjuksha@gmail.com>

Added a test case for wideangle camera lensflare

Signed-off-by: Aditya <aditya050995@gmail.com>

Add extra model in test world

The pre-existing model in the test world exhibits
the lens flare rendering bug without the fix in
this branch, so I added an extra model at a different
pose that does not exhibit the bug. It makes it
easier to confirm that the bug is fixed.

* Add extra cameras with occluded views
* Add another camera with box to its side that
  is incorrectly occluded

Signed-off-by: Steve Peters <scpeters@openrobotics.org>