Fix conda-forge GitHub Action CI by using mambaforge directly and by supporting graphviz >= 3 on Windows (#3287)
The fix is composed by two unrelated changes

### [Fix conda-forge GitHub Action CI by using mambaforge directly](https://github.com/gazebosim/gazebo-classic/commit/eea77c2d831763f6123b30e2b1a910b98205189a)

Currently the CI is failing on Linux and Windows with a weird conflict error. This is due to the fact that we are install mamba (from the conda-forge channel) on the top of miniconda3 (that by default install all the packages from the defaults channel). To make an analogy in apt world, this is like installing Debian, and then trying to install packages from the Ubuntu repo: something can go wrong.

To solve this problem, we install [mambaforge](https://github.com/conda-forge/miniforge) (a installer like miniconda3 but already using conda-forge packages) directly, so we always and only use packages from the conda-forge channel. 

Similar to:
* https://github.com/robotology/robotology-superbuild/pull/840
* https://github.com/robotology/robometry/pull/141
* https://github.com/robotology-playground/yarp-devices-ros2/pull/44

### [Add support to use graphviz >= 3 on Windows](https://github.com/gazebosim/gazebo-classic/commit/15435b1520b7fbe041cb79c5f57829ab028eb5dc)

Since graphviz 3, any downstream library that on Windows links against a graphviz shared library needs to define the `GVDLL` preprocessor definition (see https://gitlab.com/graphviz/graphviz/-/blob/3.0.0/CHANGELOG.md#changed). 

To deal with this, this PR adds the `GVDLL` definition for `gazebo_gui` when on Windows (on graphviz < 3 defining this definition will be harmless). As that prepocessor definition should not be set when linking static graphviz, and given that we can't detect in `FindGraphviz`  if the linked graphviz is static or not, an ad-hoc CMake variable `GAZEBO_FINDGRAPHVIZ_USE_STATIC_GRAPHVIZ` is introduced in the case this definition needs to be disabled. 


Similar to https://github.com/robotology/ycm/pull/414 .